refactored selectable messages to be a single group
