+++ .eslintignore
@@ -1,3 +1,2 @@
 scripts/**/*.js
 jest.config.js
-src/**/*.tsx
